owners can delete queue elements in their rooms

+++ ApiHandlers.py
@@ -158,7 +158,7 @@ class EditQueueApi(flask_restful.Resource):
 
 	def delete(self, slug, queue_id):
 		queue = Queue.objects(id=queue_id).first()
-		if current_user.is_authenticated() and queue:
+		if current_user.is_authenticated() and queue and queue.room and queue.room.owner.id==current_user.id:
 			queue.delete()
 
 
@@ -166,7 +166,7 @@ class QueueElementApi(flask_restful.Resource):
 	def delete(self, queue_id, queue_element_id):
 		queue = Queue.objects(id=queue_id).first()
 		queue_element = QueueElement.objects(id=queue_element_id).first()
-		if current_user.is_authenticated() and queue_element and current_user.id == queue_element.user.id and queue:
+		if current_user.is_authenticated() and queue_element and (current_user.id == queue_element.user.id or current_user.id == queue.room.owner.id) and queue:
 			queue.remove_queue_element(queue_element)
 			queue_element.delete()
